Use `codecs.decode` in Python 2 as opposed to `str.decode` in Python 2.
I am investigating a migration to Python 3, and to facilitate this we are using the `-3` flag as decribed here:
https://docs.python.org/3/howto/pyporting.html#prevent-compatibility-regressions .  When using this flag,
I get warnings from simplejson about the use of decode.

While obviously this is a invalid warning since this code is gated in Python 2, committing this change
will allow users to run their code in `-3` mode in Python 2 without warnings from simplejson.

Microbenchmarking this change with `timeit` also reveals it's faster!

```py
>>> import timeit
>>> timeit.timeit("decode_hex('4f6c6567')[0]", setup="import codecs; decode_hex = codecs.getdecoder('hex_codec')")
0.6572110652923584
>>> timeit.timeit("'4f6c6567'.decode('hex')")
0.9076640605926514
```