Make CLASSIFIERS a list
Prior to Python 3.7, passing a tuple classifiers (it has always
been documented as list) will return a cryptic error message
when trying to upload a release. See
https://reinbach.com/blog/setuptools-classifiers-upload-python3-5/
for example.

Python 3.7 will explicitly reject non-list types as discussed in
https://bugs.python.org/issue19610

Note that this won't break any user code if a wheel is provided
for Python 3.7.